Terminando General Summary DTO

+++ feel-flow-package/src/main/java/com/equipo5/feelflowapp/dto/dashboard/general/GeneralHappinessDto.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class GeneralHappinessDto {
     private String nameHappiness;
-    private double percentHappiness;
+    private int percentHappiness;
 }
